# Elixir: Make code more idiomatic and get x4 performance improvement