--- a/trexit/.formatter.exs
+++ b/trexit/.formatter.exs
@@ -1,4 +1,4 @@
 [
-  import_deps: [:phoenix],
+  import_deps: [:plug],
   inputs: ["*.{ex,exs}", "{config,lib,test}/**/*.{ex,exs}"]
 ]