--- a/trexit/lib/trexit/application.ex
+++ b/trexit/lib/trexit/application.ex
@@ -8,28 +8,13 @@ defmodule Trexit.Application do
   @impl true
   def start(_type, _args) do
     children = [
-      # Start the Telemetry supervisor
-      TrexitWeb.Telemetry,
-      # Start the PubSub system
-      {Phoenix.PubSub, name: Trexit.PubSub},
       Trexit.GTFS.Loader,
-      # Start the Endpoint (http/https)
-      TrexitWeb.Endpoint
-      # Start a worker by calling: Trexit.Worker.start_link(arg)
-      # {Trexit.Worker, arg}
+      {Plug.Cowboy, scheme: :http, plug: Trexit.Router, options: [port: 4000]}
     ]
 
     # See https://hexdocs.pm/elixir/Supervisor.html
     # for other strategies and supported options
     opts = [strategy: :one_for_one, name: Trexit.Supervisor]
     Supervisor.start_link(children, opts)
-  end
-
-  # Tell Phoenix to update the endpoint configuration
-  # whenever the application is updated.
-  @impl true
-  def config_change(changed, _new, removed) do
-    TrexitWeb.Endpoint.config_change(changed, removed)
-    :ok
   end
 end