--- a/trexit/lib/trexit/gtfs/stop_time.ex
+++ b/trexit/lib/trexit/gtfs/stop_time.ex
@@ -1,3 +1,0 @@
-defmodule Trexit.GTFS.StopTime do
-  defstruct [:trip_id, :stop_id, :arrival, :departure]
-end