--- a/trexit/lib/trexit/gtfs/trip.ex
+++ b/trexit/lib/trexit/gtfs/trip.ex
@@ -1,3 +1,0 @@
-defmodule Trexit.GTFS.Trip do
-  defstruct [:trip_id, :route_id, :service_id]
-end