--- a/trexit/lib/trexit/mailer.ex
+++ b/trexit/lib/trexit/mailer.ex
@@ -1,3 +1,0 @@
-defmodule Trexit.Mailer do
-  use Swoosh.Mailer, otp_app: :trexit
-end