--- a/trexit/lib/trexit_web/controllers/schedule_controller.ex
+++ b/trexit/lib/trexit_web/controllers/schedule_controller.ex
@@ -1,7 +1,0 @@
-defmodule TrexitWeb.ScheduleController do
-  use TrexitWeb, :controller
-
-  def show(conn, %{"route_id" => route_id}) do
-    json(conn, Trexit.GTFS.schedules_for_route(route_id))
-  end
-end