--- a/trexit/lib/trexit_web/templates/layout/app.html.heex
+++ b/trexit/lib/trexit_web/templates/layout/app.html.heex
@@ -1,5 +1,0 @@
-<main class="container">
-  <p class="alert alert-info" role="alert"><%= get_flash(@conn, :info) %></p>
-  <p class="alert alert-danger" role="alert"><%= get_flash(@conn, :error) %></p>
-  <%= @inner_content %>
-</main>