--- a/trexit/lib/trexit_web/views/layout_view.ex
+++ b/trexit/lib/trexit_web/views/layout_view.ex
@@ -1,7 +1,0 @@
-defmodule TrexitWeb.LayoutView do
-  use TrexitWeb, :view
-
-  # Phoenix LiveDashboard is available only in development by default,
-  # so we instruct Elixir to not warn if the dashboard route is missing.
-  @compile {:no_warn_undefined, {Routes, :live_dashboard_path, 2}}
-end