--- a/trexit/lib/trexit_web/views/page_view.ex
+++ b/trexit/lib/trexit_web/views/page_view.ex
@@ -1,3 +1,0 @@
-defmodule TrexitWeb.PageView do
-  use TrexitWeb, :view
-end