--- a/trexit/priv/static/robots.txt
+++ b/trexit/priv/static/robots.txt
@@ -1,5 +1,0 @@
-# See https://www.robotstxt.org/robotstxt.html for documentation on how to use the robots.txt file
-#
-# To ban all spiders from the entire site uncomment the next two lines:
-# User-agent: *
-# Disallow: /