--- a/trexit/test/test_helper.exs
+++ b/trexit/test/test_helper.exs
@@ -1,1 +1,0 @@
-ExUnit.start()