--- a/trexit/test/trexit_web/controllers/page_controller_test.exs
+++ b/trexit/test/trexit_web/controllers/page_controller_test.exs
@@ -1,8 +1,0 @@
-defmodule TrexitWeb.PageControllerTest do
-  use TrexitWeb.ConnCase
-
-  test "GET /", %{conn: conn} do
-    conn = get(conn, "/")
-    assert html_response(conn, 200) =~ "Welcome to Phoenix!"
-  end
-end