--- a/trexit/test/trexit_web/views/layout_view_test.exs
+++ b/trexit/test/trexit_web/views/layout_view_test.exs
@@ -1,8 +1,0 @@
-defmodule TrexitWeb.LayoutViewTest do
-  use TrexitWeb.ConnCase, async: true
-
-  # When testing helpers, you may want to import Phoenix.HTML and
-  # use functions such as safe_to_string() to convert the helper
-  # result into an HTML string.
-  # import Phoenix.HTML
-end