--- a/trexit/test/trexit_web/views/page_view_test.exs
+++ b/trexit/test/trexit_web/views/page_view_test.exs
@@ -1,3 +1,0 @@
-defmodule TrexitWeb.PageViewTest do
-  use TrexitWeb.ConnCase, async: true
-end