… trexit/assets/css/app.css
… trexit/assets/css/phoenix.css
… trexit/assets/js/app.js
… trexit/assets/vendor/topbar.js
… trexit/config/config.exs
… trexit/config/dev.exs
… trexit/config/prod.exs
… trexit/config/runtime.exs
… trexit/config/test.exs
… trexit/lib/trexit_web.ex
… trexit/lib/trexit_web/endpoint.ex
… trexit/lib/trexit_web/gettext.ex
… trexit/lib/trexit_web/router.ex
… trexit/lib/trexit_web/telemetry.ex
… trexit/lib/trexit_web/templates/layout/live.html.heex
… trexit/lib/trexit_web/templates/layout/root.html.heex
… trexit/lib/trexit_web/templates/page/index.html.heex
… trexit/lib/trexit_web/views/error_helpers.ex
… trexit/lib/trexit_web/views/error_view.ex
… trexit/mix.lock
… trexit/priv/gettext/en/LC_MESSAGES/errors.po
… trexit/priv/gettext/errors.pot
… trexit/priv/static/favicon.ico
… trexit/priv/static/images/phoenix.png
… trexit/test/support/conn_case.ex
… trexit/test/trexit_web/views/error_view_test.exs